pyproject.toml
[project]
name = "curriculum-builder"
version = "0.0.1"
description = "AI-powered personalized learning platform using Google ADK"
requires-python = ">=3.11, <3.14"
dependencies = [
    "google-adk==1.27.1",
    "google-cloud-aiplatform[adk,agent_engines]",
    "fastapi>=0.135.1",
    "uvicorn>=0.42.0",
    "python-dotenv>=1.2.2",
    "jinja2>=3.1.6",
]

classifiers = [
    "Private :: Do Not Upload",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
